remove data structure files
