+++ API/controllers/users.js
@@ -1,44 +1,30 @@
 //Import statments
-import storage from '../data/database';
 import model from '../model/user-model';
-
-/**
- * Adds an Order to the database
- * @param {*} req - incomming json data
- * @param {*} res - response to the validity of the data
- * @param {*} id  - orderId associated with the data
- */
-let pushUser = (req, res, id) => {
-    let newOrder = model.populate(req, id);
-    storage.userData.push(newOrder);
-    return newOrder;
-}
-
-/**
- * Updates an Order in the database
- * @param {*} req   - incomming json data
- * @param {*} res   - response to the validity of the data
- * @param {*} index - the index of the order being Updated
- * @param {*} id    - orderId associated with the data
- */
-let putUser = (req, index, id) => {
-    let newOrder = model.populate(req, id);
-    storage.userData[index] = newOrder;
-    return newOrder;
-}
+import database from '../db/Index'
 
 export default class Controller {
     /**
-        * Gets All Users in the database and sends as response
+        * Gets All user_accounts in the database and sends as response
         * @param {*} req - incomming request data
         * @param {*} res - response to the validity of the data 
     */
-    getUsers(req, res) {
-        res.status(200).send({
-            success: 'true',
-            status: 'Users retrieved successfully',
-            users_data: storage.userData
-        });
+    async getUsers(req, res) {
+        const command = 'SELECT * FROM user_accounts';
+        try {
+            const { rows, rowCount } = await database.query(command);
+            return res.status(200).send({
+                success: 'true',
+                status: 'Food Items retrieved successfully',
+                user_accounts: rows,
+                total_user_accounts: rowCount
+            });
+        } catch (error) {
+            return res.status(400).send({
+                success: 'false',
+                status: 'Bad Request',
+                message: 'Users Not Found'
+            });
+        }
     }
 
     /**
@@ -46,97 +32,115 @@ export default class Controller {
          * @param {*} req - incomming Request data
          * @param {*} res - response to the validity of the data
     */
-    getUser(req, res) {
-        const id = parseInt(req.params.userId, 10);
-        let userIndex;
-        storage.userData.map((user, index) => {
-            if (user.userId == id) {
-                userIndex = index;
-                let value = storage.userData[index];
-                res.status(200).send({
-                    success: 'true',
-                    status: 'User retrieved successfully',
-                    user_data: value
+    async getUser(req, res) {
+        const command = 'SELECT * FROM user_accounts WHERE user_id=$1';
+        try {
+            const { rows } = await database.query(command, [req.params.userId]);
+            if (!rows[0]) {
+                return res.status(404).send({
+                    success: 'false',
+                    status: 'User Not Found in the Database'
                 });
-            };
-        });
-        if (userIndex == undefined) {
-            return res.status(404).send({
+            } else return res.status(200).send({
+                success: 'true',
+                status: 'User retrieved successfully',
+                User: rows[0]
+            });
+        } catch (error) {
+            return res.status(400).send({
                 success: 'false',
-                Status: 'User Not Found in the Database'
-            })
+                status: 'Bad Request',
+                message: 'User Not Found'
+            });
         }
     }
 
     /**
-        * Add an user to existing Users in the database
+        * Add an User to existing user_accounts in the database
         *  @param {*} req - incomming json data
         *  @param {*} res - response to the sucess of the event
     */
-    addUser(req, res) {
-        let id = storage.userData.length;
-        let sentuser = pushUser(req, res, id);
-        res.status(201).send({
-            userId: id,
-            userData_sent: sentuser,
-            status: 'User Sent Successfully'
-        });
+    async addUser(req, res) {
+        let newUser = model.populate(req);
+        const command = `INSERT INTO
+    user_accounts(user_name,user_role,user_email,user_password)
+      VALUES($1, $2, $3, $4) returning *`;
+        try {
+            const { rows } = await database.query(command, newUser);
+            return res.status(201).send({
+                User_sent: rows[0],
+                status: 'User Sent Successfully'
+            });
+        } catch (error) {
+            return res.status(400).send({
+                success: 'false',
+                status: 'Bad Request',
+                message: error
+            });
+        }
     }
 
     /**
-     * Update an user in the database
+     * Update an User in the database
      *  @param {*} req - incomming json data
      * @param {*} res - response to the success of the event 
      */
-    updateUser(req, res) {
-
-        let map = 0;
-        const id = parseInt(req.params.userId, 10);
-        let olduser = storage.userData[id];
-        storage.userData.map((user, index) => {
-            map++;
-            if (user.userId == id) {
-                map--;
-                let state = putUser(req, index, id);
-                res.status(200).send({
-                    userId: id,
-                    old_user: olduser,
-                    user_update: state,
-                    status: "Update successful"
+    async updateUser(req, res) {
+        let user = model.populate(req);
+        let date = new Date();
+        user.push(date);
+        user.push(req.params.userId);
+        const findQuery = `SELECT * FROM user_accounts WHERE user_id=$1`;
+        const updateQuery = `UPDATE user_accounts SET user_name=$1,user_role=$2,user_email=$3,user_password=$4, modified_date=$5 WHERE user_id=$6 returning *`;
+        try {
+            const { rows } = await database.query(findQuery, [req.params.userId]);
+            if (!rows[0]) {
+                return res.status(410).send({
+                    success: 'false',
+                    status: 'Requested resourse is no longer available'
                 });
-            } else {
-                if (map == storage.userData.length) {
-                    return res.status(410).send({
-                        success: 'false',
-                        status: 'Requested User is no longer available'
-                    });
-                } else return;
             }
-        });
+            const response = await database.query(updateQuery, user);
+            return res.status(200).send({
+                userId: req.params.userId,
+                old_user: rows[0],
+                update: response.rows[0],
+                status: "Update successful"
+            });
+        } catch (error) {
+            console.log(error);
+            return res.status(400).send({
+                success: 'false',
+                status: 'Bad Request',
+                message: 'User Not Found'
+            });
+        }
     }
 
     /**
-    * Delete an user in the database
+    * Delete an User in the database
     *  @param {*} req - incomming request data
     * @param {*} res - response to the validity of the data
     */
-    deleteUser(req, res) {
-        const id = parseInt(req.params.userId, 10)
-        let userIndex;
-        storage.userData.map((user, index) => {
-            if (user.userId == id) {
-                userIndex = index;
-                storage.userData.splice(index, 1)
-                res.status(200).send({
-                    success: 'true',
-                    status: 'User deleted successfuly'
+    async deleteUser(req, res) {
+        const deleteQuery = 'DELETE FROM user_accounts WHERE user_id=$1 returning *';
+        try {
+            const { rows } = await database.query(deleteQuery, [req.params.userId]);
+            if (!rows[0]) {
+                return res.status(404).send({
+                    success: 'false',
+                    status: 'User Not Found in the Database'
                 });
             }
-        });
-        if (userIndex == undefined) {
-            return res.status(404).send({
+            return res.status(200).send({
+                success: 'true',
+                status: 'User deleted successfuly'
+            });
+        } catch (error) {
+            return res.status(400).send({
                 success: 'false',
-                Status: 'User Not Found in the Database'
+                status: 'Bad Request',
+                message: "User Not Found"
             });
         }
     }
