--- API/data/database.js
@@ -1,50 +1,0 @@
-export default {
-    database: [
-        {
-            orderId: '0',
-            food: 'peperony pizza',
-            price: '10',
-            quantity: '10',
-            orderStatus: 'uncompleted',
-            userAddress: 'Address'
-        },
-        {
-            orderId: '1',
-            food: 'grilled Chicken',
-            price: '20',
-            quantity: '20',
-            orderStatus: 'uncompleted',
-            userAddress: 'Address'
-        }
-    ],
-    userData: [
-        {
-            userId: "0",
-            userName: "pat",
-            userEmail: "pat@gmail.com",
-            userPassword: "123pat"
-        },
-        {
-            userId: "1",
-            userName: "rose",
-            userEmail: "rose@gmail.com",
-            userPassword: "123rose"
-        }
-    ],
-    foodItemsData: [
-        {
-            itemId: '0',
-            itemImage: 'imageSrc',
-            itemName: 'doughnut',
-            itemPrice: '20',
-            itemtag: 'snacks'
-        },
-        {
-            itemId: '1',
-            itemImage: 'imageSrc',
-            itemName: 'Amala',
-            itemPrice: '200',
-            itemTag: 'local dishes'
-        }
-    ]
-}
