+++ API/db/Index.js
@@ -7,7 +7,6 @@ const pool = new Pool({
     connectionString: process.env.DATABASE_URL
 });
 
-
 export default {
 /**
  * Database Query
