+++ API/middleware/foodItemsValidation.js
@@ -1,16 +1,16 @@
 /**
-* Validate post and Put request contains all required parameters.
+* Validate post and Put request contains all required parameters
 * @param {*} req - incomming API Request
 * @param {*} res - response to the validity of the data
 */
 function validation(req, res, next) {
     let error = {};
     let numbers = /[0-9]|[0-9][0-9]/g;
-    if ((req.body.itemPrice) != undefined){if (!((req.body.itemPrice).match(numbers))) {return res.status(400).send({status: 'unsuccessful', Error_Log: 'Bad Request, Price is not Valid'});}}
-    if (!(req.body.itemName)) {error.Name_Error = 'Food Name is Required';}
-    if (!(req.body.itemPrice)) {error.Price_Error ='Item Price is Required';}
-    if (!(req.body.itemTag)) {error.Tag_Error = 'Item Tag is Required';}
-    if (!(req.body.itemImage)) {error.Tag_Error = 'Item Tagz is Required';}
+    if ((req.body.item_price) != undefined){if (!((req.body.item_price).match(numbers))) {return res.status(400).send({status: 'unsuccessful', Error_Log: 'Bad Request, Price is not Valid'});}}
+    if (!(req.body.item_name)) {error.Name_Error = 'Item Name is Required';}
+    if (!(req.body.item_price)) {error.Price_Error ='Item Price is Required';}
+    if (!(req.body.item_tag)) {error.Tag_Error = 'Item Tag is Required';}
+    if (!(req.body.item_image)) {error.Tag_Error = 'Item Image is Required';}
     if ((Object.keys(error).length) > 0){ return res.status(400).send({status: 'Bad Request', success:"false",  Error_Log: error});}
     next();
 }
