+++ package.json
@@ -13,7 +13,7 @@
     "clear": "rimraf dist coverage",
     "coverage": "./node_modules/.bin/istanbul cover ./node_modules/mocha/bin/_mocha -- -R spec ./dist/API/test/*.js --exit",
     "create": "babel-node ./db.js",
-    "heroku-postbuild": "babel-node ./dist/db.js"
+    "heroku-postbuild": "babel-node ./db.js"
   },
   "repository": {
     "type": "git",
